+++ lib/src/buttons/animated/animated_block_level_button.dart
@@ -2,36 +2,35 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:modular_ui/src/utils/dimensions.dart';
 
-/// A customizable gradient block level button by ModularUI with optional icons.
-class MUIGradientBlockButton extends StatefulWidget {
-  const MUIGradientBlockButton({
-    super.key,
-    required this.text,
-    required this.onPressed,
-    this.bgGradient = const LinearGradient(colors: [Colors.black, Colors.grey]),
-    this.textColor = Colors.white,
-    this.borderRadius = 10,
-    this.hapticsEnabled = false,
-    this.animationDuration = 250,
-    this.widthFactorPressed = 0.95,
-    this.heightFactor = 0.05,
-    this.leadingIcon,
-    this.actionIcon,
-    this.iconColor = Colors.white,
-    this.widthFactor = 0.98,
-    this.boxShadows,
-  });
+/// A customizable animated button which takes the complete available width
+class MUIAnimatedBlockButton extends StatefulWidget {
+  const MUIAnimatedBlockButton(
+      {super.key,
+      required this.text,
+      required this.onPressed,
+      this.bgColor = Colors.black,
+      this.textStyle = const TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+      this.borderRadius = 10,
+      this.hapticsEnabled = false,
+      this.animationDuration = 250,
+      this.widthFactorPressed = 0.95,
+      this.heightFactor = 0.05,
+      this.leadingIcon,
+      this.actionIcon,
+      this.iconColor = Colors.white,
+      this.widthFactor = 0.98,
+      this.boxShadows});
 
   /// The Text to display inside the button
   final String text;
 
-  /// Background Gradient of The Gradient Block Button, default: Black and grey
-  final Gradient bgGradient;
+  /// Background Color of The Animated Block Button, default: Black
+  final Color bgColor;
 
-  /// Text Color of the Gradient Block Button, default: white
-  final Color textColor;
+  /// Text style of the Animated Block Button, default: white
+  final TextStyle textStyle;
 
-  /// Border Radius for Gradient Block Button, default: 10
+  /// Border Radius for Animated Block Button, default: 10
   final double borderRadius;
 
   /// Animation Duration in Milliseconds, default: 250 ms
@@ -65,18 +64,18 @@ class MUIGradientBlockButton extends StatefulWidget {
   final List<BoxShadow>? boxShadows;
 
   @override
-  State<MUIGradientBlockButton> createState() => _MUIGradientBlockButtonState();
+  State<MUIAnimatedBlockButton> createState() => _MUIAnimatedBlockButtonState();
 }
 
-class _MUIGradientBlockButtonState extends State<MUIGradientBlockButton> {
-  bool _isGradientBlockLevelButtonPressed = false;
+class _MUIAnimatedBlockButtonState extends State<MUIAnimatedBlockButton> {
+  bool _isAnimatedBlockLevelButtonPressed = false;
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onTapDown: (_) {
         setState(() {
-          _isGradientBlockLevelButtonPressed = true;
+          _isAnimatedBlockLevelButtonPressed = true;
         });
         if (widget.hapticsEnabled) {
           HapticFeedback.lightImpact();
@@ -84,24 +83,23 @@ class _MUIGradientBlockButtonState extends State<MUIGradientBlockButton> {
       },
       onTapUp: (_) {
         setState(() {
-          _isGradientBlockLevelButtonPressed = false;
+          _isAnimatedBlockLevelButtonPressed = false;
         });
         widget.onPressed();
       },
       onTapCancel: () {
         setState(() {
-          _isGradientBlockLevelButtonPressed = false;
+          _isAnimatedBlockLevelButtonPressed = false;
         });
       },
       child: AnimatedContainer(
         duration: Duration(milliseconds: widget.animationDuration),
         decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(widget.borderRadius),
-            gradient: widget.bgGradient,
-            boxShadow: widget.boxShadows),
-        width: _isGradientBlockLevelButtonPressed
-            ? getScreenWidth(context) * widget.widthFactorPressed
-            : getScreenWidth(context) * widget.widthFactor,
+          borderRadius: BorderRadius.circular(widget.borderRadius),
+          color: widget.bgColor,
+          boxShadow: widget.boxShadows,
+        ),
+        width: _isAnimatedBlockLevelButtonPressed ? getScreenWidth(context) * widget.widthFactorPressed : getScreenWidth(context) * widget.widthFactor,
         height: getScreenHeight(context) * widget.heightFactor,
         child: Center(
           child: Row(
@@ -115,13 +113,7 @@ class _MUIGradientBlockButtonState extends State<MUIGradientBlockButton> {
                   size: 12,
                 ),
               SizedBox(width: widget.leadingIcon != null ? 8.0 : 0.0),
-              Text(
-                widget.text,
-                style: TextStyle(
-                  color: widget.textColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
+              Text(widget.text, style: widget.textStyle),
               SizedBox(width: widget.actionIcon != null ? 8.0 : 0.0),
               if (widget.actionIcon != null)
                 Icon(
