+++ lib/src/buttons/outlined/outlined_button.dart
@@ -1,18 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:modular_ui/src/utils/dimensions.dart';
 
-class MUISecondaryOutlinedButton extends StatefulWidget {
-  const MUISecondaryOutlinedButton({
+class MUIOutlinedButton extends StatefulWidget {
+  const MUIOutlinedButton({
     super.key,
     required this.text,
     required this.onPressed,
-    required this.bgColor,
-    this.borderColor = Colors.grey,
-    this.borderWidth = 1.5,
-    this.textColor = Colors.black,
+    this.bgColor = Colors.transparent,
+    this.borderColor = Colors.black,
+    this.borderWidth = 2,
+    this.textStyle = const TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
     this.borderRadius = 10,
     this.hapticsEnabled = false,
     this.animationDuration = 250,
+    this.widthFactorUnpressed = 0.04,
+    this.widthFactorPressed = 0.035,
+    this.heightFactorUnPressed = 0.03,
+    this.heightFactorPressed = 0.03,
     this.maxHorizontalPadding = 50,
     this.leadingIcon,
     this.actionIcon,
@@ -23,19 +28,19 @@ class MUISecondaryOutlinedButton extends StatefulWidget {
   /// The Text to display inside the button
   final String text;
 
-  /// Text Color of the Secondary Outlined Button, default : black
-  final Color textColor;
-
-  /// Border Color for the outlined Button also equals to background color when button is tapped
-  final Color borderColor;
-
-  /// Background color of secondary outlined button
+  /// Background Color of The Outlined Button, default : transparent
   final Color bgColor;
 
-  /// Border width, default : 1.5
+  /// Text Style of the Outlined Button, default : black
+  final TextStyle textStyle;
+
+  /// Border Color for the outlined Button
+  final Color borderColor;
+
+  /// Border width, default : 2
   final double borderWidth;
 
-  /// Border Radius for Secondary Outlined Button, default : 10
+  /// Border Radius for Outlined Button, default : 10
   final double borderRadius;
 
   /// Animation Duration in Milliseconds, default : 250 ms
@@ -43,6 +48,18 @@ class MUISecondaryOutlinedButton extends StatefulWidget {
 
   /// Enables Light Haptic Feedback
   final bool hapticsEnabled;
+
+  /// A double value which gets multiplied by the current screen width when button is not pressed
+  final double widthFactorUnpressed;
+
+  /// A double value which gets multiplied by the current screen width when button is pressed
+  final double widthFactorPressed;
+
+  /// A double value which gets multiplied by the current screen height when button is  pressed
+  final double heightFactorPressed;
+
+  /// A double value which gets multiplied by the current screen height when button is not pressed
+  final double heightFactorUnPressed;
 
   /// A double value which determines maximum horizontal padding a button can accumulate
   /// Play with this value if you want to use the button on a larger screen size
@@ -64,12 +81,10 @@ class MUISecondaryOutlinedButton extends StatefulWidget {
   final List<BoxShadow>? boxShadows;
 
   @override
-  State<MUISecondaryOutlinedButton> createState() =>
-      _MUISecondaryOutlinedButtonState();
+  State<MUIOutlinedButton> createState() => _MUIOutlinedButtonState();
 }
 
-class _MUISecondaryOutlinedButtonState
-    extends State<MUISecondaryOutlinedButton> {
+class _MUIOutlinedButtonState extends State<MUIOutlinedButton> {
   bool _isOutlinedButtonPressed = false;
   @override
   Widget build(BuildContext context) {
@@ -97,14 +112,26 @@ class _MUISecondaryOutlinedButtonState
         duration: Duration(milliseconds: widget.animationDuration),
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(widget.borderRadius),
-          color: _isOutlinedButtonPressed ? widget.borderColor : widget.bgColor,
+          color: widget.bgColor,
           boxShadow: widget.boxShadows,
           border: Border.all(
-            color: widget.borderColor,
+            color: _isOutlinedButtonPressed ? Colors.transparent : widget.borderColor,
             width: widget.borderWidth,
           ),
         ),
-        padding: const EdgeInsets.symmetric(horizontal: 18, vertical: 10),
+        padding: EdgeInsets.symmetric(
+          horizontal: _isOutlinedButtonPressed ? getScreenWidth(context) * widget.widthFactorPressed : getScreenWidth(context) * widget.widthFactorUnpressed,
+          vertical: _isOutlinedButtonPressed ? getScreenWidth(context) * widget.heightFactorPressed : getScreenWidth(context) * widget.heightFactorUnPressed,
+        ).clamp(
+          const EdgeInsets.symmetric(
+            horizontal: 10,
+            vertical: 8,
+          ),
+          EdgeInsets.symmetric(
+            horizontal: widget.maxHorizontalPadding,
+            vertical: 16,
+          ),
+        ),
         child: Row(
           mainAxisSize: MainAxisSize.min,
           children: [
@@ -116,11 +143,7 @@ class _MUISecondaryOutlinedButtonState
                 size: 12,
               ),
             SizedBox(width: widget.leadingIcon != null ? 8.0 : 0.0),
-            Text(
-              widget.text,
-              style: TextStyle(
-                  color: widget.textColor, fontWeight: FontWeight.bold),
-            ),
+            Text(widget.text, style: widget.textStyle),
             SizedBox(width: widget.actionIcon != null ? 8.0 : 0.0),
             if (widget.actionIcon != null)
               Icon(
