+++ lib/src/buttons/secondary/secondary_block_level_button.dart
@@ -2,34 +2,47 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:modular_ui/src/utils/dimensions.dart';
 
-/// A customizable text block level button by ModularUI with optional icons.
-class MUITextBlockButton extends StatefulWidget {
-  const MUITextBlockButton({
+/// A customizable secondary block level button by ModularUI with optional icons.
+class MUISecondaryBlockButton extends StatefulWidget {
+  const MUISecondaryBlockButton({
     super.key,
     required this.text,
     required this.onPressed,
-    this.bgColor = Colors.grey,
-    this.textColor = Colors.black,
+    this.bgColor = Colors.transparent,
+    this.textStyle = const TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
+    this.borderColor = Colors.grey,
+    this.tappedBgColor = Colors.grey,
+    this.borderWidth = 2,
     this.borderRadius = 10,
     this.hapticsEnabled = false,
     this.animationDuration = 250,
-    this.widthFactor = 0.95,
     this.heightFactor = 0.05,
     this.leadingIcon,
     this.actionIcon,
     this.iconColor = Colors.black,
+    this.widthFactor = 0.98,
+    this.boxShadows,
   });
 
   /// The Text to display inside the button
   final String text;
 
-  /// Background Color of The Text Block Button is pressed, default: grey
+  /// Background Color of The Block Button
   final Color bgColor;
 
-  /// Text Color of the Text Block Button, default: black
-  final Color textColor;
+  /// Background Color of The Block Button
+  final Color tappedBgColor;
 
-  /// Border Radius for Text Block Button, default: 10
+  /// Text Style of the  Block Button
+  final TextStyle textStyle;
+
+  /// Border Color for the Button
+  final Color borderColor;
+
+  /// Border width, default: 2
+  final double borderWidth;
+
+  /// Border Radius for Block Button, default: 10
   final double borderRadius;
 
   /// Animation Duration in Milliseconds, default: 250 ms
@@ -38,11 +51,14 @@ class MUITextBlockButton extends StatefulWidget {
   /// Enables Light Haptic Feedback
   final bool hapticsEnabled;
 
-  /// A double value which gets multiplied by the current screen width determines the width of the button in pressedState
+  /// A double value which gets multiplied by the current screen height determines the height of the button
+  final double heightFactor;
+
+  /// A double value which gets multiplied with current screen width when button is not pressed
   final double widthFactor;
 
-  /// A double value which gets multiplied by the current screen height determines the height of the button in pressedState
-  final double heightFactor;
+  /// On Pressed Function
+  final VoidCallback onPressed;
 
   /// Optional leading icon for the button.
   final IconData? leadingIcon;
@@ -50,25 +66,25 @@ class MUITextBlockButton extends StatefulWidget {
   /// Optional action icon for the button.
   final IconData? actionIcon;
 
-  /// Icon color for both leading and action icons, default: white.
+  /// Icon color for both leading and action icons, default: black.
   final Color iconColor;
 
-  /// On Pressed Function
-  final VoidCallback onPressed;
+  /// Box shadows for button
+  final List<BoxShadow>? boxShadows;
 
   @override
-  State<MUITextBlockButton> createState() => _MUITextBlockButtonState();
+  State<MUISecondaryBlockButton> createState() => _MUISecondaryBlockButtonState();
 }
 
-class _MUITextBlockButtonState extends State<MUITextBlockButton> {
-  bool _isTextBlockLevelButtonPressed = false;
+class _MUISecondaryBlockButtonState extends State<MUISecondaryBlockButton> {
+  bool _isSecondaryBlockLevelButtonPressed = false;
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onTapDown: (_) {
         setState(() {
-          _isTextBlockLevelButtonPressed = true;
+          _isSecondaryBlockLevelButtonPressed = true;
         });
         if (widget.hapticsEnabled) {
           HapticFeedback.lightImpact();
@@ -76,22 +92,25 @@ class _MUITextBlockButtonState extends State<MUITextBlockButton> {
       },
       onTapUp: (_) {
         setState(() {
-          _isTextBlockLevelButtonPressed = false;
+          _isSecondaryBlockLevelButtonPressed = false;
         });
         widget.onPressed();
       },
       onTapCancel: () {
         setState(() {
-          _isTextBlockLevelButtonPressed = false;
+          _isSecondaryBlockLevelButtonPressed = false;
         });
       },
       child: AnimatedContainer(
         duration: Duration(milliseconds: widget.animationDuration),
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(widget.borderRadius),
-          color: _isTextBlockLevelButtonPressed
-              ? widget.bgColor
-              : Colors.transparent,
+          color: _isSecondaryBlockLevelButtonPressed ? widget.tappedBgColor : widget.bgColor,
+          boxShadow: widget.boxShadows,
+          border: Border.all(
+            color: _isSecondaryBlockLevelButtonPressed ? widget.tappedBgColor : widget.borderColor,
+            width: widget.borderWidth,
+          ),
         ),
         width: getScreenWidth(context) * widget.widthFactor,
         height: getScreenHeight(context) * widget.heightFactor,
@@ -109,10 +128,7 @@ class _MUITextBlockButtonState extends State<MUITextBlockButton> {
               SizedBox(width: widget.leadingIcon != null ? 8.0 : 0.0),
               Text(
                 widget.text,
-                style: TextStyle(
-                  color: widget.textColor,
-                  fontWeight: FontWeight.bold,
-                ),
+                style: widget.textStyle
               ),
               SizedBox(width: widget.actionIcon != null ? 8.0 : 0.0),
               if (widget.actionIcon != null)
