+++ lib/src/buttons/text/text_block_level_button.dart
@@ -2,44 +2,34 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:modular_ui/src/utils/dimensions.dart';
 
-/// A customizable outlined block level button by ModularUI with optional icons.
-class MUIOutlinedBlockButton extends StatefulWidget {
-  const MUIOutlinedBlockButton({
+/// A customizable text block level button by ModularUI with optional icons.
+class MUITextBlockButton extends StatefulWidget {
+  const MUITextBlockButton({
     super.key,
     required this.text,
     required this.onPressed,
-    this.bgColor = Colors.transparent,
-    this.textColor = Colors.black,
-    this.borderColor = Colors.black,
-    this.borderWidth = 2,
+    this.bgColor = Colors.grey,
+    this.textStyle = const TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
     this.borderRadius = 10,
     this.hapticsEnabled = false,
     this.animationDuration = 250,
-    this.widthFactorPressed = 0.95,
+    this.widthFactor = 0.95,
     this.heightFactor = 0.05,
     this.leadingIcon,
     this.actionIcon,
     this.iconColor = Colors.black,
-    this.widthFactor = 0.98,
-    this.boxShadows,
   });
 
   /// The Text to display inside the button
   final String text;
 
-  /// Background Color of The Outlined Block Button, default: transparent
+  /// Background Color of The Text Block Button is pressed, default: grey
   final Color bgColor;
 
-  /// Text Color of the Outlined Block Button, default: black
-  final Color textColor;
+  /// Text Style of the Text Block Button, default: black
+  final TextStyle textStyle;
 
-  /// Border Color for the outlined Button
-  final Color borderColor;
-
-  /// Border width, default: 2
-  final double borderWidth;
-
-  /// Border Radius for Outlined Block Button, default: 10
+  /// Border Radius for Text Block Button, default: 10
   final double borderRadius;
 
   /// Animation Duration in Milliseconds, default: 250 ms
@@ -48,17 +38,11 @@ class MUIOutlinedBlockButton extends StatefulWidget {
   /// Enables Light Haptic Feedback
   final bool hapticsEnabled;
 
-  /// A double value which gets multiplied by the current screen width when the button is pressed
-  final double widthFactorPressed;
-
-  /// A double value which gets multiplied by the current screen height determines the height of the button
-  final double heightFactor;
-
-  /// A double value which gets multiplied with current screen width when button is not pressed
+  /// A double value which gets multiplied by the current screen width determines the width of the button in pressedState
   final double widthFactor;
 
-  /// On Pressed Function
-  final VoidCallback onPressed;
+  /// A double value which gets multiplied by the current screen height determines the height of the button in pressedState
+  final double heightFactor;
 
   /// Optional leading icon for the button.
   final IconData? leadingIcon;
@@ -66,25 +50,25 @@ class MUIOutlinedBlockButton extends StatefulWidget {
   /// Optional action icon for the button.
   final IconData? actionIcon;
 
-  /// Icon color for both leading and action icons, default: black.
+  /// Icon color for both leading and action icons, default: white.
   final Color iconColor;
 
-  /// Box shadows for button
-  final List<BoxShadow>? boxShadows;
+  /// On Pressed Function
+  final VoidCallback onPressed;
 
   @override
-  State<MUIOutlinedBlockButton> createState() => _MUIOutlinedBlockButtonState();
+  State<MUITextBlockButton> createState() => _MUITextBlockButtonState();
 }
 
-class _MUIOutlinedBlockButtonState extends State<MUIOutlinedBlockButton> {
-  bool _isOutlinedBlockLevelButtonPressed = false;
+class _MUITextBlockButtonState extends State<MUITextBlockButton> {
+  bool _isTextBlockLevelButtonPressed = false;
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onTapDown: (_) {
         setState(() {
-          _isOutlinedBlockLevelButtonPressed = true;
+          _isTextBlockLevelButtonPressed = true;
         });
         if (widget.hapticsEnabled) {
           HapticFeedback.lightImpact();
@@ -92,29 +76,22 @@ class _MUIOutlinedBlockButtonState extends State<MUIOutlinedBlockButton> {
       },
       onTapUp: (_) {
         setState(() {
-          _isOutlinedBlockLevelButtonPressed = false;
+          _isTextBlockLevelButtonPressed = false;
         });
         widget.onPressed();
       },
       onTapCancel: () {
         setState(() {
-          _isOutlinedBlockLevelButtonPressed = false;
+          _isTextBlockLevelButtonPressed = false;
         });
       },
       child: AnimatedContainer(
         duration: Duration(milliseconds: widget.animationDuration),
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(widget.borderRadius),
-          color: widget.bgColor,
-          boxShadow: widget.boxShadows,
-          border: Border.all(
-            color: widget.borderColor,
-            width: widget.borderWidth,
-          ),
+          color: _isTextBlockLevelButtonPressed ? widget.bgColor : Colors.transparent,
         ),
-        width: _isOutlinedBlockLevelButtonPressed
-            ? getScreenWidth(context) * widget.widthFactorPressed
-            : getScreenWidth(context) * widget.widthFactor,
+        width: getScreenWidth(context) * widget.widthFactor,
         height: getScreenHeight(context) * widget.heightFactor,
         child: Center(
           child: Row(
@@ -128,13 +105,7 @@ class _MUIOutlinedBlockButtonState extends State<MUIOutlinedBlockButton> {
                   size: 12,
                 ),
               SizedBox(width: widget.leadingIcon != null ? 8.0 : 0.0),
-              Text(
-                widget.text,
-                style: TextStyle(
-                  color: widget.textColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
+              Text(widget.text, style: widget.textStyle),
               SizedBox(width: widget.actionIcon != null ? 8.0 : 0.0),
               if (widget.actionIcon != null)
                 Icon(
